combine pv/pvc protection controllers
