+++ pkg/controller/BUILD
@@ -134,8 +134,7 @@ filegroup(
         "//pkg/controller/volume/events:all-srcs",
         "//pkg/controller/volume/expand:all-srcs",
         "//pkg/controller/volume/persistentvolume:all-srcs",
-        "//pkg/controller/volume/pvcprotection:all-srcs",
-        "//pkg/controller/volume/pvprotection:all-srcs",
+        "//pkg/controller/volume/storageprotection:all-srcs",
     ],
     tags = ["automanaged"],
 )
